requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.4
python-multipart==0.0.9
SQLAlchemy==2.0.29
psycopg2-binary==2.9.9
GeoAlchemy2==0.14.7
python-dotenv==1.0.1
firebase-admin==6.5.0
websockets==12.0
boto3==1.34.64
gunicorn==21.2.0